encapsulated geodetic conversion code into a separate directory. Added FIXME for future work.
