+++ search/lib/convert/geodeticCoordinates/coordinate.js
@@ -1,11 +1,26 @@
 const DEG_TO_RAD = Math.PI / 180;
 const RAD_TO_DEG = 180 / Math.PI;
 
+// export interface LatLng{
+//   lat: number;
+//   lng: number;
+// }
+
 // Class for dealing with conversions between lat/lng, phi/theta, and x/y/z as well
 // as operations on the various forms.
 // Consider properties on this class to be immutable.  Changing, say, 'x' will not
 // update `phi` or `theta` and will throw normalization out of whack.
 export class Coordinate {
+  // readonly phi: number;
+
+  // readonly theta: number;
+
+  // readonly x: number;
+
+  // readonly y: number;
+
+  // readonly z: number;
+
   constructor (phi, theta, x, y, z) {
     this.phi = phi;
     this.theta = theta;
@@ -14,11 +29,11 @@ export class Coordinate {
     this.z = z;
   }
 
-  static fromLngLat (...args) {
+  static fromLatLng (...args) {
     let lat;
     let lng;
     if (args.length === 1) {
-      [{ lng, lat }] = (args);
+      [{ lat, lng }] = (args);
     } else {
       [lng, lat] = (args);
     }
