+++ search/lib/convert/granules.js
@@ -2,7 +2,6 @@ const _ = require('lodash');
 const cmr = require('../cmr');
 const settings = require('../settings');
 const { pointStringToPoints, parseOrdinateString, addPointsToBbox, mergeBoxes, reorderBoxValues } = require('./bounding-box');
-const { inflectBox } = require('./geodeticCoordinates');
 const { generateAppUrl, generateAppUrlWithoutRelativeRoot, wfs, extractParam, generateSelfUrl } = require('../util');
 
 function cmrPolygonToGeoJsonPolygon (polygon) {
@@ -78,14 +77,10 @@ function cmrSpatialToGeoJSONGeometry (cmrGran) {
 function cmrSpatialToStacBbox (cmrGran) {
   let bbox = null;
   if (cmrGran.polygons) {
-    // bbox = cmrGran.polygons
-    // .map((rings) => rings[0])
-    // .map(pointStringToPoints);
-    // .reduce(addPointsToBbox, bbox);
-    const polygonPoints = cmrGran.polygons.map((rings) => rings[0])
-      .map(pointStringToPoints);
-    console.log(polygonPoints);
-    bbox = inflectBox(polygonPoints[0]);
+    bbox = cmrGran.polygons
+      .map((rings) => rings[0])
+      .map(pointStringToPoints)
+      .reduce(addPointsToBbox, bbox);
   }
   if (cmrGran.points) {
     const points = cmrGran.points.map(parseOrdinateString);
